+++ pypeit/scripts/ql.py
@@ -52,19 +52,21 @@ from pypeit import msgs
 from pypeit import pypeitsetup
 from pypeit import metadata
 from pypeit import io
+from pypeit import inputfiles 
 from pypeit import pypeit
+from pypeit import coadd2d
 from pypeit.par.pypeitpar import PypeItPar
 from pypeit.calibframe import CalibFrame
 from pypeit.core.parse import parse_binning
 from pypeit.scripts import scriptbase
 from pypeit.spectrographs import available_spectrographs
 from pypeit.slittrace import SlitTraceSet 
-from pypeit import inputfiles 
 
 from pypeit.scripts.setup_coadd2d import SetupCoAdd2D
 from pypeit.scripts.coadd_2dspec import CoAdd2DSpec
+from pypeit.scripts.show_2dspec import Show2DSpec
 
-def get_files(raw_files, raw_path, ext):
+def get_files(raw_files, raw_path):
     """
     Use the user-provided input to get the files to process.
 
@@ -75,9 +77,6 @@ def get_files(raw_files, raw_path, ext):
         raw_path (:obj:`str`):
             The path to the raw files parsed from the ``raw_path`` command line
             argument.
-        ext (:obj:`str`):
-            The file extension used to search for raw files parsed from the
-            ``ext`` command line argument.
 
     Returns:
         :obj:`list`: List of strings providing the full path to each raw file.
@@ -93,11 +92,10 @@ def get_files(raw_files, raw_path, ext):
             pass
     if files is None: 
         try:
-            files = inputfiles.grab_rawfiles(raw_paths=[raw_path], list_of_files=raw_files,
-                                             extension=ext) 
+            files = inputfiles.grab_rawfiles(raw_paths=[raw_path], list_of_files=raw_files)
         except PypeItError as e:
-            msgs.error('Unable to parse provided input files.  Check --raw_files, --raw_path, '
-                        'and/or --ext input.')
+            msgs.error('Unable to parse provided input files.  Check --raw_files and '
+                       '--raw_path input.')
     return files
 
 
@@ -156,19 +154,51 @@ def quicklook_regroup(fitstbl):
             if 'target' in fitstbl.keys() \
                     and not all(fitstbl['target'][is_type] == fitstbl['target'][is_type][0]):
                 msgs.error(f'All {frametype} frames must be of the same target.')
+
+            # Regroup dithered observations so that all images at a unique
+            # offset are combined.
             if 'bkg_id' in fitstbl.keys() and any(fitstbl['bkg_id'].data[is_type] != -1):
                 if 'dithoff' not in fitstbl.keys():
                     msgs.error('CODING ERROR: Metadata does not include dithoff column!')
                 # Group the unique dither positions
                 dith, inv = np.unique(fitstbl['dithoff'].data[is_type], return_inverse=True)
-                fitstbl['comb_id'][is_type] = inv + comb_strt
-                # NOTE: This does the "dumb" thing of just setting the background
-                # image for the first offset to the second one, and setting the
-                # background for all other offsets to be the first one.
-                fitstbl['bkg_id'][np.where(is_type)[0][inv == comb_strt]] = comb_strt+1
-                fitstbl['bkg_id'][np.where(is_type)[0][inv != comb_strt]] = comb_strt
+                if len(dith) == 1:
+                    msgs.warn('All exposures have the same offset!')
+                    fitstbl['comb_id'][is_type] = comb_strt
+                else:
+                    # This creates comb+bkg pairs that match the absolute value of the offset
+                    # - Find the unique absolute values
+                    abs_dith, abs_inv = np.unique(np.absolute(dith), return_inverse=True)
+                    # - Create the indices that will identify the background images
+                    bkg_inv = np.arange(dith.size)
+                    # - For each unique throw, swap the combination IDs for the background IDs
+                    for i in range(abs_dith.size):
+                        bkg_inv[abs_inv == i] = bkg_inv[abs_inv == i][::-1]
+                    # - Get the background indices
+                    bkg_inv = bkg_inv[inv]
+
+                    if np.array_equal(bkg_inv, inv):
+                        # All of the offsets are unique!  Do the "dumb" thing of
+                        # just setting the background image for the first offset
+                        # to the second one, and setting the background for all
+                        # other offsets to be the first one.
+                        fitstbl['bkg_id'][np.where(is_type)[0][inv == comb_strt]] = comb_strt+1
+                        fitstbl['bkg_id'][np.where(is_type)[0][inv != comb_strt]] = comb_strt
+                    else:
+                        same_frame = np.equal(bkg_inv, inv)
+                        if any(same_frame):
+                            # One or more of the offsets has no matched
+                            # absolute-value pair.  Just use one of the other
+                            # background images.
+                            indx = np.where(np.logical_not(same_frame))[0][-1]
+                            bkg_inv[same_frame] = bkg_inv[indx]
+                        # All the offsets have a matched absolute value pair
+                        fitstbl['comb_id'][is_type] = inv + comb_strt
+                        fitstbl['bkg_id'][is_type] = bkg_inv + comb_strt
             else:
+                # Force all images to be combined
                 fitstbl['comb_id'][is_type] = comb_strt
+            # Prep for the next image type (only if there are standards!)
             comb_strt = np.amax(fitstbl['comb_id'].data[is_type])+1
 
 
@@ -176,7 +206,7 @@ def generate_sci_pypeitfile(redux_path:str,
                             ref_calib_dir:Path,
                             ps_sci, 
                             det:str=None,
-                            clean:bool=False,
+                            clear:bool=False,
                             slitspatnum:str=None,
                             maskID:str=None,
                             boxcar_radius:float=None,
@@ -203,7 +233,7 @@ def generate_sci_pypeitfile(redux_path:str,
             Setup object for the science frame(s) only.
         det (:obj:`str`, optional):
             Detector/mosaic identifier.  If None, all detectors are reduced.
-        clean (:obj:`bool`, optional):
+        clear (:obj:`bool`, optional):
             Remove the directory structure for these files; i.e., start a
             completely clean reduction.  If false, any existing directory
             structure will remain, but any existing science reduction products
@@ -241,7 +271,7 @@ def generate_sci_pypeitfile(redux_path:str,
     calib_dir = sci_dir / ps_sci.par['calibrations']['calib_dir']
 
     # Science reduction folder
-    if sci_dir.exists() and clean:
+    if sci_dir.exists() and clear:
         shutil.rmtree(sci_dir)
     # NOTE: This is done here (as opposed to waiting for the pypeit.PypeIt call
     # to create it) so that the symlink to the calibrations directory can be
@@ -264,7 +294,7 @@ def generate_sci_pypeitfile(redux_path:str,
     # the reduced standard and remove the standard from the setup object.
     std_spec1d = None
     is_std = ps_sci.fitstbl.find_frames('standard', index=True)
-    if len(is_std) > 0 and not clean:
+    if len(is_std) > 0 and not clear:
         for i in is_std:
             std_spec1d = pypeit.PypeIt.get_spec_file_name(
                             str(sci_dir / ps_sci.par['rdx']['scidir']),
@@ -278,7 +308,7 @@ def generate_sci_pypeitfile(redux_path:str,
             # NOTE: Should not need to regroup!
             msgs.warn(f'Found existing standard star reduction: {std_spec1d}.  This will be used '
                       'and the standards will not be re-reduced!  To force them to be '
-                      're-reduced, use the --clean option.')
+                      're-reduced, use the --clear_science option.')
             ps_sci.remove_table_rows(is_std)
 
     # TODO: Push this stuff into a function in par/pypeitpar.py?
@@ -397,7 +427,7 @@ def calib_manifest(calib_dir, spectrograph):
     # Check the calibration directory exists
     _calib_dir = Path(calib_dir).resolve()
     if not _calib_dir.exists():
-        msgs.error(f'Calibration directory does not exist: {_calib_dir}')
+        return None
 
     # Find the pypeit files
     pypeit_files = sorted(_calib_dir.glob('{0}_*/{0}*.pypeit'.format(spectrograph)))
@@ -450,10 +480,14 @@ def match_to_calibs(ps:pypeitsetup.PypeItSetup, calib_dir:str, calibrated_setups
         If a setup in ``ps`` is not matched, the associated dictionary item is
         None.
     """
-    if calibrated_setups is None:
-        calibrated_setups = calib_manifest(calib_dir, ps.spectrograph.name)
-
     matched_configs = {}
+    if calibrated_setups is None:
+        # Try to build the calibration manifest
+        calibrated_setups = calib_manifest(calib_dir, ps.spectrograph.name)
+    if calibrated_setups is None:
+        # If it still doesn't exist, return empty dictionary
+        return matched_configs
+
     for setup, config in ps.fitstbl.configs.items():
         if calibrated_setups is None:
             matched_configs[setup] = None
@@ -525,7 +559,7 @@ def merge_setups(calibrated_setups, calib_match, frame_setup):
     return new_frame_setup.tolist()
 
 
-def get_setup_calib(calib_dir):
+def get_setup_calib(calib_dir, calib_grp=None):
     """
     Using all of the files found in the provided directory, determine the setup
     and calibration group to use for the quicklook science frame(s).
@@ -533,6 +567,10 @@ def get_setup_calib(calib_dir):
     Args:
         calib_dir (:obj:`str`, `Path`_):
             Directory with the calibration files.
+        calib_grp (:obj:`int`, optional):
+            If the calibration directory contains results from more than one
+            calibration group, this *must* be provided, specifying which
+            calibration group should be used.
 
     Returns:
         :obj:`tuple`: The setup name and calibration group.
@@ -575,11 +613,14 @@ def get_setup_calib(calib_dir):
     if len(unique_calibs) == 1:
         # There's only one calibration group, so use it.
         return setup, unique_calibs[0]
+    if calib_grp is not None:
+        if str(calib_grp) in unique_calibs:
+            return setup, str(calib_grp)
+        msgs.error(f'Selected calibration group {calib_grp} is not available in {_calib_dir}.  '
+                   'Must select a valid group.  Directory currently contains the following '
+                   f'calibration groups: {unique_calibs}')
 
-    # TODO: For now, we fault here.  For quick-look, this means there should
-    # only be one calibration group, or all the calibrations are assigned to
-    # 'all' calibration groups.  Otherwise, we could have the user supply the
-    # calibration group.
+    # Cannot determine which calibration group to use.
     msgs.error(f'Calibrations in {_calib_dir} are part of multiple calibration groups.  Unclear '
                'how to proceed.')
 
@@ -603,17 +644,14 @@ class QL(scriptbase.ScriptBase):
         parser.add_argument('--raw_path', type=str, default='current working directory',
                             help='Directory with the raw files to process.  Ignored if a '
                                  'PypeIt-formatted file is provided using the --rawfiles option.')
-        parser.add_argument('--ext', type=str, default='.fits',
-                            help='If raw file names are not provided directly using the '
-                                 '--rawfiles option, this sets the extension used when searching '
-                                 'for any files in the path defined by --raw_path.  All files '
-                                 'found in the raw path with this extension will be processed.')
+        # TODO: Allow user to specify target to be reduced?
 
         parser.add_argument('--sci_files', type=str, nargs='+',
                             help='A space-separated list of raw file names that are science '
                                  'exposures.  These files must *also* be in the list of raw '
                                  'files.  Use of this option overrides the automated PypeIt '
-                                 'frame typing.')
+                                 'frame typing.  Should only be used of automatic frame typing '
+                                 'fails or is undesirable.')
 
         parser.add_argument('--redux_path', type=str, default='current working directory',
                             help='Path for the QL reduction outputs.')
@@ -632,27 +670,22 @@ class QL(scriptbase.ScriptBase):
                                  'in that directory that match the instrument setup/configuration '
                                  'of the provided data, the code will construct new calibrations '
                                  '(assuming relevant raw files are provided).')
-        parser.add_argument('--clean', default=False, action='store_true',
-                            help='Remove the existing output directories to force a fresh '
+        parser.add_argument('--clear_science', default=False, action='store_true',
+                            help='Remove the existing output science directories to force a fresh '
                                  'reduction.  If False, any existing directory structure will '
-                                 'remain, but any existing science files will still be '
-                                 'overwritten.')
+                                 'remain, and any alterations to existing science files will '
+                                 'follow the normal behavior of run_pypeit.')
 
-        # TODO: Options to add?:
-        #   - do not overwrite existing reductions
-        #   - specify `target` to be reduced, either parsed from the frame
-        #     header (i.e., target in the fitstbl) or from the mask information
-
+        # TODO: Get rid of "calibs only"?  I.e., force user only provide
+        # calibrations if they want to process only calibrations.  Maybe the
+        # point is that the latter isn't clear for frames typed as
+        # `arc,science,tilt`...
         parser.add_argument('--calibs_only', default=False, action='store_true',
                             help='Reduce only the calibrations?')
+
         parser.add_argument('--overwrite_calibs', default=False, action='store_true',
-                            help='Overwrite any existing calibration files?')
-        parser.add_argument('--slitspatnum', type=str,
-                            help='Reduce the slit(s) as specified by the slitspatnum value(s)')
-        parser.add_argument('--maskID', type=int,
-                            help='Reduce the slit(s) as specified by the maskID value(s)')
-        parser.add_argument('--boxcar_radius', type=float,
-                            help='Set the radius for the boxcar extraction in arcseconds')
+                            help='Re-process and overwrite any existing calibration files.')
+
         parser.add_argument('--det', type=str, nargs='+',
                             help='A space-separated set of detectors or detector mosaics to '
                                  'reduce.  By default, *all* detectors or default mosaics for '
@@ -662,18 +695,49 @@ class QL(scriptbase.ScriptBase):
                                  'separately reduce detectors 1 and 5 for Keck/DEIMOS, you would '
                                  'use --det 1 5; to reduce mosaics made up of detectors 1,5 and '
                                  '3,7, you would use --det 1,5 3,7')
+        parser.add_argument('--slitspatnum', type=str,
+                            help='Reduce the slit(s) as specified by the slitspatnum value(s)')
+        parser.add_argument('--maskID', type=int,
+                            help='Reduce the slit(s) as specified by the maskID value(s)')
+        parser.add_argument('--boxcar_radius', type=float,
+                            help='Set the radius for the boxcar extraction in arcseconds')
+        parser.add_argument('--snr_thresh', default=None, type=float,
+                            help='Change the default S/N threshold used during source detection')
+
         parser.add_argument('--ignore_std', default=False, action='store_true',
                             help='If standard star observations are automatically detected, '
                                  'ignore those frames.  Otherwise, they are included with the '
                                  'reduction of the science frames.')
-        parser.add_argument('--snr_thresh', default=None, type=float,
-                            help='Change the default S/N threshold used during source detection')
+        parser.add_argument('--skip_display', dest='show', default=True, action='store_false',
+                            help='Run the quicklook without displaying any results.')
 
-        parser.add_argument('--coadd', default=False, action='store_true',
+        # TODO: Add fluxing option?
+
+        # Coadding options
+        parser.add_argument('--coadd2d', default=False, action='store_true',
                             help='Perform default 2D coadding.')
-        # TODO: Add in relevant 2d coadding parameters, like detector, slits,
-        # objects, etc.  2D coadding is basically just a default run of
-        # `pypeit_setup_coadd2d` and `pypeit_coadd_2dspec`.
+        # TODO: Consolidate slitspatnum and only_slits!
+        parser.add_argument('--only_slits', type=str, nargs='+',
+                            help='If coadding, only coadd this space-separated set of slits.  If '
+                                 'not provided, all slits are coadded.')
+        parser.add_argument('--offsets', type=str, default=None,
+                            help='If coadding, spatial offsets to apply to each image; see the '
+                                 '[coadd2d][offsets] parameter.  Options are restricted here to '
+                                 'either maskdef_offsets or auto.  If not specified, the '
+                                 '(spectrograph-specific) default is used.')
+        parser.add_argument('--weights', type=str, default=None,
+                            help='If coadding, weights used to coadd images; see the '
+                                 '[coadd2d][weights] parameter.  Options are restricted here to '
+                                 'either uniform or auto.  If not specified, the '
+                                 '(spectrograph-specific) default is used.')
+        parser.add_argument('--spec_samp_fact', default=1.0, type=float,
+                            help='If coadding, adjust the wavelength grid sampling by this '
+                                 'factor.  For a finer grid, set value to <1.0; for coarser '
+                                 'sampling, set value to >1.0).')
+        parser.add_argument('--spat_samp_fact', default=1.0, type=float,
+                            help='If coadding, adjust the spatial grid sampling by this '
+                                 'factor.  For a finer grid, set value to <1.0; for coarser '
+                                 'sampling, set value to >1.0).')
 
         return parser
 
@@ -683,12 +747,10 @@ class QL(scriptbase.ScriptBase):
 
         tstart = time.perf_counter()
 
-        #assert False
-
         # Parse the raw files
-        files = get_files(args.raw_files, args.raw_path, args.ext)
+        files = get_files(args.raw_files, args.raw_path)
         if len(files) == 0:
-            msgs.error('No files to read!  Check --raw_files, --raw_path, and/or --ext input.')
+            msgs.error('No files to read!  Check --raw_files and --raw_path input.')
 
         # TODO: Include an option to save the ingested file list as a PypeIt
         # RawFile that can be edited?
@@ -701,20 +763,26 @@ class QL(scriptbase.ScriptBase):
         ps = pypeitsetup.PypeItSetup.from_rawfiles(files, args.spectrograph)
         ps.run(setup_only=True)
 
-        # Check for any untyped files
-        unknown_types = [t is None for t in ps.fitstbl['frametype']]
-        if any(unknown_types):
-            # TODO: Remove them and keep going or fault?
-            # TODO: Check these against ones that have been specified as science using 'sci_files'
-            msgs.error('Could not determine frame types for the following files: ' +
-                       ', '.join(ps.fitstbl['frametype'][unknown_types]))
-
         # Find the raw science files
         sci_idx = ps.fitstbl.find_frames('science') if args.sci_files is None \
                         else np.in1d(ps.fitstbl['filename'].data, args.sci_files)
+        # TODO: Allow for standard files to be identified?
+
+        # Check for any untyped files (that have not been typed) as science
+        # files.  NOTE: This works regardless of whether or not sci_files has
+        # been defined.
+        unknown_types = [t is None for t in ps.fitstbl['frametype']]
+        if any(unknown_types & np.logical_not(sci_idx)):
+            # TODO: Remove them and keep going instead?
+            msgs.error('Could not determine frame types for the following files: ' +
+                       ', '.join(ps.fitstbl['filename'][unknown_types & np.logical_not(sci_idx)]))
+
+        # Include any standards? 
+        # TODO: Standards are only reduced if there are also science frames?
         if any(sci_idx) and not args.ignore_std:
-            # Include standards
-            sci_idx |= ps.fitstbl.find_frames('standard')
+            std_idx = ps.fitstbl.find_frames('standard')
+        else:
+            std_idx = np.zeros(sci_idx.size, dtype=bool)
 
         # Set the directory with the calibrations
         setup_calib_dir = None if args.setup_calib_dir is None \
@@ -725,13 +793,25 @@ class QL(scriptbase.ScriptBase):
         if any(sci_idx):
             # Generate science setup object
             sci_files = ps.fitstbl.frame_paths(sci_idx)
-            ps_sci = pypeitsetup.PypeItSetup.from_rawfiles(sci_files, ps.spectrograph.name)
+            use_type = 'arc,science,tilt' \
+                            if 'arc' in ps.fitstbl['frametype'][sci_idx][0] else 'science'
+            frametype = {Path(f).name : use_type for f in sci_files}
+            if any(std_idx):
+                std_files = ps.fitstbl.frame_paths(std_idx)
+                use_type = 'arc,standard,tilt' \
+                                if 'arc' in ps.fitstbl['frametype'][std_idx][0] else 'standard'
+                frametype = {**frametype, **{Path(f).name : use_type for f in std_files}}
+                sci_files += std_files
+            ps_sci = pypeitsetup.PypeItSetup.from_rawfiles(sci_files, ps.spectrograph.name,
+                                                           frametype=frametype)
             # NOTE: By default, groupings=True in PypeItSetup.run, meaning that
             # the combination groups should be automatically set.  Setting the
             # combination groups is needed when standards are included and to
             # automatically identify dithered observations for some
             # spectrographs.
             ps_sci.run(setup_only=True)
+
+            # TODO: Ensure that the file types are correct
 
             # Limit to a single setup
             if len(ps_sci.fitstbl.configs.keys()) > 1:
@@ -771,7 +851,7 @@ class QL(scriptbase.ScriptBase):
 
             # Check that all the frames are assigned to the same calibration group
             # NOTE: This is largely superfluous given the use of get_setup_calib
-            # in generate_sci_pyepitfile, but it's useful to keep the warning
+            # in generate_sci_pypeitfile, but it's useful to keep the warning
             # here.
             if any(ps_sci.fitstbl['calib'] != ps_sci.fitstbl['calib'][0]):
                 msgs.warn('Automated configuration assigned multiple calibration groups to your '
@@ -783,8 +863,12 @@ class QL(scriptbase.ScriptBase):
                 # relevant set of existing calibrations
                 setup_calib_dir = match_to_calibs(ps_sci, args.parent_calib_dir)
                 if setup_calib_dir is None:
+                    # TODO: Fault here, or keep going to the next step, which is
+                    # to try to build the calibrations?
                     msgs.error('No calibrations exist or could not find appropriate setup match '
                                f'in provided parent directory: {args.parent_calib_dir}')
+                # NOTE: Code above check that there is only one setup in ps_sci
+                setup_calib_dir = setup_calib_dir[ps_sci.fitstbl['setup'][0]]['calib_dir']
                 msgs.info(f'Attempting to use archived calibrations found in {setup_calib_dir}.')
 
         elif not args.calibs_only:
@@ -801,6 +885,7 @@ class QL(scriptbase.ScriptBase):
             # Set the parent directory
             parent_calib_dir = args.redux_path if args.parent_calib_dir is None \
                                     else args.parent_calib_dir
+
             # Find all of the existing calibrations for this setup in the parent
             # directory
             calibrated_setups = calib_manifest(parent_calib_dir, ps.spectrograph.name)
@@ -851,6 +936,7 @@ class QL(scriptbase.ScriptBase):
                 # Run
                 pypeIt = pypeit.PypeIt(calib_pypeit_file,
                                        redux_path=str(redux_path),
+                                       reuse_calibs=not args.overwrite_calibs,
                                        calib_only=True)
                 pypeIt.calib_all()
 
@@ -863,18 +949,13 @@ class QL(scriptbase.ScriptBase):
         sci_pypeit_file = generate_sci_pypeitfile(
                     args.redux_path, 
                     setup_calib_dir, 
-                    ps_sci, 
+                    ps_sci,
                     det=args.det,
-                    clean=args.clean,
+                    clear=args.clear_science,
                     slitspatnum=args.slitspatnum,
                     maskID=args.maskID, 
                     boxcar_radius=args.boxcar_radius,
                     snr_thresh=args.snr_thresh)
-
-#        from pypeit.inputfiles import PypeItFile
-#        f = PypeItFile.from_file(sci_pypeit_file)
-#        embed()
-#        exit()
 
         # Run it
         pypeIt = pypeit.PypeIt(sci_pypeit_file, reuse_calibs=True)
@@ -883,34 +964,57 @@ class QL(scriptbase.ScriptBase):
 
         # TODO: There are currently no tests for this!!
         # Perform coadding if requested
-        if args.coadd:
+        if args.coadd2d:
             # Run the setup script to get the baseline coadd2d file
+            # NOTE: By running the setup script with ``--keep_par`` flag,
+            # parameters that skip the sky subtraction and the optimal
+            # extraction should be propagated to the 2D coadding...
             # TODO:
-            #   - Add sensitivity function
-            #   - Allow SetupCoAdd2D to auto-detect dithers based on the pypeit
-            #     file and parse/prepare the coadding offsets accordingly
-            #   - and other options
-            SetupCoAdd2D.main(SetupCoAdd2D.parse_args([sci_pypeit_file]))
+            #   - Add sensitivity function?
+            command_line_args = ['-f', sci_pypeit_file, '--keep_par']
+            if args.only_slits is not None:
+                command_line_args += ['--only_slits'] + args.only_slits
+            if args.offsets is not None:
+                command_line_args += ['--offsets', args.offsets]
+            if args.weights is not None:
+                command_line_args += ['--weights', args.weights]
+            SetupCoAdd2D.main(SetupCoAdd2D.parse_args(command_line_args))
 
             # Find all the coadd2d scripts
             # NOTE: This should only find *one* coadd2d file because quick-look
             # should be limited to performing reduction for one target at a
             # time.
-            # TODO: Need to have SetupCoAdd2D.main return the names of the
-            # written files...
-            coadd_files = sorted(Path(sci_pypeit_file).resolve().parent.glob('*.coadd2d'))
+            coadd_file = sorted(Path(sci_pypeit_file).resolve().parent.glob('*.coadd2d'))
+            if len(coadd_file) != 1:
+                msgs.error('There should be only one 2D coadd file.')
+            coadd_file = coadd_file[0]
             
-            # Run the coadding, only on those coadd files with more than one file
-            for coadd_file in coadd_files:
-                coadd2dFile = inputfiles.Coadd2DFile.from_file(coadd_file)
-                if len(coadd2dFile.data) < 2:
-                    msgs.warn(f'{coadd_file} only has one spec2d file.  Continuing...')
-                    continue
+            # Run the coadding
+            coadd2dFile = inputfiles.Coadd2DFile.from_file(coadd_file)
+            CoAdd2DSpec.main(CoAdd2DSpec.parse_args([str(coadd_file),
+                                                     '--spec_samp_fact', str(args.spec_samp_fact),
+                                                     '--spat_samp_fact', str(args.spat_samp_fact)]))
 
-                # TODO: Add options (e.g. spatial/spectral sampling...)
-                CoAdd2DSpec.main(CoAdd2DSpec.parse_args([coadd_file]))
+            # Get the output file name
+            spectrograph, par, _ = coadd2dFile.get_pypeitpar()
+            spec2d_files = coadd2dFile.filenames
+            coadd_scidir = Path(coadd2d.CoAdd2D.output_paths(spec2d_files, par)[0]).resolve()
+            basename = coadd2d.CoAdd2D.default_basename(spec2d_files)
+            spec2d_file = str(coadd_scidir / f'spec2d_{basename}.fits')
+        else:
+            # Grab the spec2d file (or at least the first one)
+            frame = pypeIt.fitstbl.find_frames('science', index=True)[0]
+            spec2d_file = pypeIt.spec_output_file(frame, twod=True)
 
-        # TODO: Add some show options
+        if args.show:
+            # TODO: Need to parse detector here?
+            Show2DSpec.main(Show2DSpec.parse_args([spec2d_file]))
+
+        # TODO: 
+        #   - Print a statement that allows users to copy-paste the correct
+        #     pypeit_show_1dspec call?
+        #   - Provide a summary of the objects (that's not buried in previous
+        #     screen output)?
 
         exec_s = np.around(time.perf_counter()-tstart, decimals=1)
         msgs.info(f'Quicklook execution time: {datetime.timedelta(seconds=exec_s)}')